+++ core-ng/src/main/java/core/framework/impl/scheduler/WeeklyTrigger.java
@@ -3,24 +3,26 @@ package core.framework.impl.scheduler;
 import core.framework.api.scheduler.Job;
 
 import java.time.DayOfWeek;
-import java.time.Duration;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 
 /**
- * @author tminglei
+ * @author tminglei, neo
  */
-public final class WeeklyTrigger implements Trigger {
+public final class WeeklyTrigger implements DynamicTrigger {
     private final String name;
     private final Job job;
     private final DayOfWeek dayOfWeek;
     private final LocalTime time;
+    private final ZoneId zoneId;
 
-    public WeeklyTrigger(String name, Job job, DayOfWeek dayOfWeek, LocalTime time) {
+    public WeeklyTrigger(String name, Job job, DayOfWeek dayOfWeek, LocalTime time, ZoneId zoneId) {
         this.name = name;
         this.job = job;
         this.dayOfWeek = dayOfWeek;
         this.time = time;
+        this.zoneId = zoneId;
     }
 
     @Override
@@ -35,8 +37,8 @@ public final class WeeklyTrigger implements Trigger {
 
     @Override
     public void schedule(Scheduler scheduler) {
-        Duration delay = nextDelay(LocalDateTime.now());
-        scheduler.schedule(this, delay, Duration.ofDays(7));
+        ZonedDateTime next = next(ZonedDateTime.now());
+        scheduler.schedule(this, next);
     }
 
     @Override
@@ -44,13 +46,13 @@ public final class WeeklyTrigger implements Trigger {
         return "weekly@" + dayOfWeek + "/" + time;
     }
 
-    Duration nextDelay(LocalDateTime now) {
-        LocalDateTime target = LocalDateTime.of(now.toLocalDate(), time).plusDays(dayOfWeek.getValue() - now.getDayOfWeek().getValue());
-
-        Duration delay = Duration.between(now, target);
-        if (delay.isNegative()) {
-            return delay.plusDays(7);
+    @Override
+    public ZonedDateTime next(ZonedDateTime now) {
+        ZonedDateTime targetZonedDateTime = now.withZoneSameInstant(zoneId);
+        ZonedDateTime next = targetZonedDateTime.plusDays(dayOfWeek.getValue() - now.getDayOfWeek().getValue()).with(time);
+        if (!next.isAfter(now)) {
+            next = next.plusWeeks(1).with(time);     // reset time in case the current day is daylight saving start date
         }
-        return delay;
+        return next;
     }
 }